requirements/base.txt
django==1.9.0
pillow==3.0.0
django-object-actions==0.5.1
tweetpony==1.5.2
psycopg2==2.6.1
python3-memcached==1.51
